T83 Successfully implement recursive user tree
Succesfully implemented the recursive user tree population. This works
for any levels deep of sub-tasks. It does take a while to complete
though because each population is executed individually. It would be
nice to be able to do it all in one request to the MongoDB, but it seems
a bit complicated at the moment getting lookup or graphLookup to work
right. Given the time constraints, this method seemed necessary.

In the future it would be nice to have a single lookup that performs the
sweep of tasks or something, or maybe a graphLookup, then assemble those
aka "marshalling" into the completed UserDoc.