Add Menu Bar and basic Task display
The App Bar element is the basis of the menu but im not sure how useful
it will be. Im still partial on a tab system.

Implementing a Left-hand side bar next. Probably going to put the menu
bar in the side bar and create a button for a setting display. (Maybe
take over the side bar?).

Projects will go in the side bar and highlight the one thats selected.
That should make it clear to the user what project their working on
and make navigation easier.