US96T104 Remove eslint extends declaration
This was causing a lot of problems because it seems that when an eslint
config is specifed in the package.json it overrides anything specified
in the .eslintrc.json file. This reveals some very hard to debug
problems and makes it so the eslint config needs to be kept in two
places. This was removed to make it so the config can just be managed
out of the .eslintrc.json file.