Add TextField elements to task table
Order Now! supplies are limited!

The TextField is basic with no error checking but thats all i could do
with the couple hours i had.

I was also implementing a change counter to limit the amount of changes
sent to the server. Im sure you have another way to do it but thats what
I got for now.