US121 Create AppSaveStatus and ClientData folder
Created an AppSaveStatus class that holds on to the current saved status
of the application and provides methods to add listeners to it. This
made it so that the app didn't entirely re-render whenever a change was
made.

Also moved ClientData into a folder to try and keep some logical
consistency. Could definitely be a better solution. But couldn't think
of a better one at the time.