T150 Add utils folder, reorganize components
Added index.tsx to files that were based around one component and had
their child components in them. Also renamed the big ClientData class to
UserData because that is more along the line of what it was. Renamed
logic folder to utils because that seems to be a common practice among
large react projects.