Fixed bugs in PrereqTaskManager
A ) Modified the PrereqProjectTaskList to be able to add projects to the
prereq tasks.

B ) Added a new List for selected prereq tasks (AllPrereqTasks) to allow
separation of projects and tasks into 2. Allowing for better clarity of
the underlying functionality.

C ) Tried to fix some of the styling to pretty it up... it didnt work. I
may need to rewrite wome of the JSX.

D ) Added a sorting function to allow the projects to be sorted by
prerequisites.

E ) Other smaller refactors.